+++ lib/improver/utilities/cube_manipulation.py
@@ -38,17 +38,8 @@ import numpy as np
 import iris
 from iris.coords import AuxCoord, DimCoord
 from iris.exceptions import CoordinateNotFoundError
+from improver.constants import MODEL_ID_DICT
 from improver.utilities.cube_checker import check_cube_coordinates
-
-# Define model_id keys to match mosg__model_configuration values
-MODEL_ID_DICT = {
-    1000: 'gl_det',
-    2000: 'gl_ens',
-    3000: 'uk_det',
-    4000: 'uk_ens',
-    5000: 'nc_det',
-    6000: 'nc_ens'
-    }
 
 
 def _associate_any_coordinate_with_master_coordinate(
